Move Checkbox to design system
