+++ web/src/features/blobstorage-integration/components/ExportFieldGroupsField.tsx
@@ -1,5 +1,5 @@
 import { useFormState, useWatch } from "react-hook-form";
-import { Checkbox } from "@/src/components/ui/checkbox";
+import { Checkbox } from "@/src/components/design-system/Checkbox/Checkbox";
 import {
   FormDescription,
   FormField,
