+++ web/src/features/models/components/pricing-tiers/TierConditionsEditor.tsx
@@ -2,7 +2,7 @@ import { PlusCircle, Trash2 } from "lucide-react";
 import { useFieldArray } from "react-hook-form";
 import { Button } from "@/src/components/ui/button";
 import { Input } from "@/src/components/ui/input";
-import { Checkbox } from "@/src/components/ui/checkbox";
+import { Checkbox } from "@/src/components/design-system/Checkbox/Checkbox";
 import {
   FormControl,
   FormDescription,
